Renamed to custom_command.py
